Separate admin-facing vs user-facing javascript
