+++ app/assets/javascripts/spotlight/admin/add_new_page_button.js
