+++ app/assets/javascripts/spotlight/admin/appearance.js
