+++ app/assets/javascripts/spotlight/admin/attachments.js
