+++ app/assets/javascripts/spotlight/admin/blacklight_configuration.js
