+++ app/assets/javascripts/spotlight/admin/block_mixins/autocompleteable.js
