+++ app/assets/javascripts/spotlight/admin/block_mixins/formable.js
