+++ app/assets/javascripts/spotlight/admin/block_mixins/plustextable.js
