+++ app/assets/javascripts/spotlight/admin/blocks/block.js
