+++ app/assets/javascripts/spotlight/admin/blocks/iframe_block.js
