+++ app/assets/javascripts/spotlight/admin/blocks/oembed_block.js
