+++ app/assets/javascripts/spotlight/admin/blocks/resources_block.js
