+++ app/assets/javascripts/spotlight/admin/blocks/rule_block.js
