+++ app/assets/javascripts/spotlight/admin/blocks/uploaded_items_block.js
