+++ app/assets/javascripts/spotlight/admin/catalog_edit.js
