+++ app/assets/javascripts/spotlight/admin/copy_email_addresses.js
