+++ app/assets/javascripts/spotlight/admin/crop.es6
