+++ app/assets/javascripts/spotlight/admin/edit_in_place.js
