+++ app/assets/javascripts/spotlight/admin/exhibits.js
