+++ app/assets/javascripts/spotlight/admin/form_observer.js
