+++ app/assets/javascripts/spotlight/admin/iiif.es6
