+++ app/assets/javascripts/spotlight/admin/index.js
@@ -1,0 +1,18 @@
+//= require underscore
+//= require eventable
+//= require sir-trevor
+//= require nestable
+//= require parameterize
+//= require bootstrap-tagsinput
+//= require jquery.serializejson
+//= require clipboard
+//= require leaflet
+//= require leaflet-iiif
+//= require Leaflet.Editable
+//= require Path.Drag
+//= require MutationObserver
+
+// Provide AMD module support
+//= require almond
+//= require polyfill.min.js
+//= require_tree .
