+++ app/assets/javascripts/spotlight/admin/locks.js
