+++ app/assets/javascripts/spotlight/admin/multi_image_selector.js
