+++ app/assets/javascripts/spotlight/admin/pages.js.erb
@@ -37,6 +37,4 @@ Spotlight.onLoad(function(){
 
     new Spotlight.BlockLimits(editor).enforceLimits(editor);
   }
-
-  $('.carousel').carousel();
 });
