+++ app/assets/javascripts/spotlight/admin/readonly_checkbox.js
