+++ app/assets/javascripts/spotlight/admin/reindex_monitor.js
