+++ app/assets/javascripts/spotlight/admin/select_related_input.js
