+++ app/assets/javascripts/spotlight/admin/sir-trevor/block_controls.js
