+++ app/assets/javascripts/spotlight/admin/sir-trevor/locales.js
