+++ app/assets/javascripts/spotlight/admin/spotlight_nestable.js
