+++ app/assets/javascripts/spotlight/admin/tabs.js
