+++ app/assets/javascripts/spotlight/admin/translation_progress.js
