+++ app/assets/javascripts/spotlight/admin/users.js
