+++ app/assets/javascripts/spotlight/user/analytics.js
