+++ app/assets/javascripts/spotlight/user/carousel.js
@@ -1,0 +1,3 @@
+Spotlight.onLoad(function(){
+  $('.carousel').carousel();
+});
