+++ app/assets/javascripts/spotlight/user/clear_form_button.js
