+++ app/assets/javascripts/spotlight/user/index.js
@@ -1,0 +1,6 @@
+//= require spotlight/spotlight
+//= require bootstrap/tab
+//= require bootstrap/tooltip
+//= require bootstrap/popover
+//= require bootstrap/carousel
+//= require_tree .
