+++ app/assets/javascripts/spotlight/user/report_a_problem.js
