+++ app/assets/javascripts/spotlight/user/zpr_links.js.erb
